backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pytest
httpx
pandas
numpy
scikit-learn
python-jose
passlib[bcrypt]
email-validator